move setup our to test_utils
