+++ src/rust/e2e-tests/src/test_utils/context.rs
@@ -1,3 +1,4 @@
+use bytes::Bytes;
 use clap::Parser;
 use grapl_config::PostgresClient;
 use grapl_tracing::{
@@ -19,12 +20,23 @@ use rust_proto::{
         BuildGrpcClientOptions,
     },
     graplinc::grapl::api::{
-        event_source::v1beta1::client::EventSourceServiceClient,
+        event_source::v1beta1::{
+            client::EventSourceServiceClient,
+            CreateEventSourceRequest,
+        },
         pipeline_ingress::v1beta1::client::PipelineIngressClient,
-        plugin_registry::v1beta1::PluginRegistryServiceClient,
+        plugin_registry::v1beta1::{
+            DeployPluginRequest,
+            PluginMetadata,
+            PluginRegistryServiceClient,
+            PluginType,
+        },
     },
 };
 use test_context::AsyncTestContext;
+use uuid::Uuid;
+
+use crate::test_fixtures;
 
 pub struct E2eTestContext {
     pub event_source_client: EventSourceServiceClient,
@@ -78,5 +90,70 @@ impl AsyncTestContext for E2eTestContext {
             plugin_work_queue_psql_client,
             _guard,
         }
+    }
+}
+
+pub struct SetupResult {
+    pub tenant_id: Uuid,
+    pub plugin_id: Uuid,
+    pub event_source_id: Uuid,
+}
+
+impl E2eTestContext {
+    pub async fn setup_sysmon_generator(
+        &mut self,
+        test_name: &str,
+    ) -> Result<SetupResult, Box<dyn std::error::Error>> {
+        let plugin_artifact = test_fixtures::get_sysmon_generator()?;
+        self.setup_generator(test_name, plugin_artifact).await
+    }
+
+    async fn setup_generator(
+        &mut self,
+        test_name: &str,
+        plugin_artifact: Bytes,
+    ) -> Result<SetupResult, Box<dyn std::error::Error>> {
+        tracing::info!(">> Settting up Event Source, Plugin");
+
+        let tenant_id = Uuid::new_v4();
+
+        // Register an Event Source
+        let event_source = self
+            .event_source_client
+            .create_event_source(CreateEventSourceRequest {
+                display_name: test_name.to_string(),
+                description: "arbitrary".to_string(),
+                tenant_id,
+            })
+            .await?;
+
+        // Deploy a Generator Plugin that responds to that event_source_id
+        let plugin = {
+            let plugin = self
+                .plugin_registry_client
+                .create_plugin(
+                    PluginMetadata {
+                        tenant_id,
+                        display_name: test_name.to_string(),
+                        plugin_type: PluginType::Generator,
+                        event_source_id: Some(event_source.event_source_id),
+                    },
+                    futures::stream::once(async move { plugin_artifact }),
+                )
+                .await?;
+
+            self.plugin_registry_client
+                .deploy_plugin(DeployPluginRequest {
+                    plugin_id: plugin.plugin_id,
+                })
+                .await?;
+            plugin
+        };
+
+        Ok(SetupResult {
+            tenant_id,
+            plugin_id: plugin.plugin_id,
+            event_source_id: event_source.event_source_id,
+        })
     }
 }
