+++ src/rust/e2e-tests/tests/sysmon_log_e2e_test.rs
@@ -4,7 +4,10 @@ use bytes::Bytes;
 use e2e_tests::{
     test_fixtures,
     test_utils::{
-        context::E2eTestContext,
+        context::{
+            E2eTestContext,
+            SetupResult,
+        },
         predicates::{
             events_36lines_merged_graph_predicate,
             events_36lines_node_identity_predicate,
@@ -18,31 +21,23 @@ use kafka::{
 use plugin_work_queue::test_utils::scan_for_plugin_message_in_pwq;
 use rust_proto::graplinc::grapl::{
     api::{
-        event_source::v1beta1::CreateEventSourceRequest,
         graph::v1beta1::GraphDescription,
         pipeline_ingress::v1beta1::PublishRawLogRequest,
-        plugin_registry::v1beta1::{
-            DeployPluginRequest,
-            PluginMetadata,
-            PluginType,
-        },
     },
     pipeline::v1beta1::RawLog,
 };
 use test_context::test_context;
-use uuid::Uuid;
 
 #[test_context(E2eTestContext)]
 #[tokio::test]
 async fn test_sysmon_log_e2e(ctx: &mut E2eTestContext) -> Result<(), Box<dyn std::error::Error>> {
     let test_name = "test_sysmon_log_e2e";
-    let tenant_id = Uuid::new_v4();
 
     let SetupResult {
         tenant_id,
         plugin_id,
         event_source_id,
-    } = common_setup(ctx, test_name, tenant_id).await?;
+    } = ctx.setup_sysmon_generator(test_name).await?;
 
     tracing::info!(">> Setup complete. Now let's test milestones in the pipeline.");
 
@@ -109,57 +104,4 @@ async fn test_sysmon_log_e2e(ctx: &mut E2eTestContext) -> Result<(), Box<dyn std
     // TODO: Perhaps add a test here that looks in dgraph/scylla for those identified nodes
 
     Ok(())
-}
-
-pub struct SetupResult {
-    tenant_id: Uuid,
-    plugin_id: Uuid,
-    event_source_id: Uuid,
-}
-async fn common_setup(
-    ctx: &mut E2eTestContext,
-    test_name: &str,
-    tenant_id: Uuid,
-) -> Result<SetupResult, Box<dyn std::error::Error>> {
-    tracing::info!(">> Settting up Event Source, Plugin");
-
-    // Register an Event Source
-    let event_source = ctx
-        .event_source_client
-        .create_event_source(CreateEventSourceRequest {
-            display_name: test_name.to_string(),
-            description: "arbitrary".to_string(),
-            tenant_id,
-        })
-        .await?;
-
-    // Deploy a Generator Plugin that responds to that event_source_id
-    let plugin = {
-        let plugin_artifact = test_fixtures::get_sysmon_generator()?;
-        let plugin = ctx
-            .plugin_registry_client
-            .create_plugin(
-                PluginMetadata {
-                    tenant_id,
-                    display_name: test_name.to_string(),
-                    plugin_type: PluginType::Generator,
-                    event_source_id: Some(event_source.event_source_id),
-                },
-                futures::stream::once(async move { plugin_artifact }),
-            )
-            .await?;
-
-        ctx.plugin_registry_client
-            .deploy_plugin(DeployPluginRequest {
-                plugin_id: plugin.plugin_id,
-            })
-            .await?;
-        plugin
-    };
-
-    Ok(SetupResult {
-        tenant_id,
-        plugin_id: plugin.plugin_id,
-        event_source_id: event_source.event_source_id,
-    })
 }
